+++ discord/objects/__init__.py
@@ -1,3 +1,4 @@
+from . import components
 from .activity import Activity
 from .activity_assets import ActivityAssets
 from .activity_button import ActivityButton